requirements.txt
asgiref==3.4.1
dj-database-url==0.5.0
dj-static==0.0.6
Django==4.0
djangorestframework==3.12.4
docopt==0.6.2
gitignore==0.0.8
gunicorn==20.1.0
psycopg2==2.9.2
sqlparse==0.4.2
static3==0.7.0
tzdata==2021.5
whitenoise==5.3.0
